差实验3和实验4，中断进入不稳定

+++ USER/lab6.h
@@ -8,7 +8,7 @@ typedef struct
 	uint32_t cnt;
 	uint32_t period_num;
 	uint32_t period;
-	uint32_t  frequency;
+	uintmax_t  frequency;
 	uint32_t time_ms;
 	uint16_t period_uint;
 }lab6_parm_t;
